requirements.txt
auth0-python>=4.7.0
python-dotenv>=1.0.0
boto3>=1.34.0
requests>=2.31.0
google-auth>=2.29.0
google-api-python-client>=2.120.0
